Minor changes and a new tagger
Minor bug fixes to V1.2, such as the fact that passing a duration to
tone() does not also count as a delay (whoops).
The new file is a set of basic firmware for the prototype soon to be
built inside a cheap plastic "gun" housing for a play station move
controller.  (Yeah "gun" shapes are against the rules but this allows me
to test accuracy in a frame type I know and can aim very easily/well, as
well as the actual casing makes it very easy to swap test LED
configurations, so I'm using it for the prototype.)