Moar Functionality!
This code should be working. It needs the display functions added so
that it can display stuff.... Need to test with actual hardware to debug
although software tests suggest that it can receive hits, kill player,
track respawns, set new configurations.
Supposedly, it can shoot and reload. Have not tested.